+++ library/core/src/test/java/com/google/android/exoplayer2/drm/DefaultDrmSessionManagerTest.java
@@ -25,6 +25,7 @@ import androidx.annotation.Nullable;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.Format;
+import com.google.android.exoplayer2.analytics.PlayerId;
 import com.google.android.exoplayer2.drm.DrmSessionManager.DrmSessionReference;
 import com.google.android.exoplayer2.drm.ExoMediaDrm.AppManagedProvider;
 import com.google.android.exoplayer2.source.MediaSource;
@@ -67,12 +68,11 @@ public class DefaultDrmSessionManagerTest {
             .setUuidAndExoMediaDrmProvider(DRM_SCHEME_UUID, uuid -> new FakeExoMediaDrm())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
 
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED_WITH_KEYS);
@@ -91,12 +91,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
 
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED_WITH_KEYS);
@@ -117,12 +116,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
     drmSession.release(/* eventDispatcher= */ null);
 
@@ -140,12 +138,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
     drmSession.release(/* eventDispatcher= */ null);
 
@@ -165,12 +162,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED_WITH_KEYS);
 
@@ -192,12 +188,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     drmSessionManager.release();
 
     // The manager is now in a 'releasing' state because the session is still active - so the
@@ -238,12 +233,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
 
     // Release the manager (there's still an explicit reference to the session from acquireSession).
@@ -278,12 +272,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession firstDrmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(firstDrmSession);
     firstDrmSession.release(/* eventDispatcher= */ null);
 
@@ -293,9 +286,7 @@ public class DefaultDrmSessionManagerTest {
     DrmSession secondDrmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                secondFormatWithDrmInitData));
+                /* eventDispatcher= */ null, secondFormatWithDrmInitData));
     // The drmSessionManager had to release firstDrmSession in order to acquire secondDrmSession.
     assertThat(firstDrmSession.getState()).isEqualTo(DrmSession.STATE_RELEASED);
 
@@ -322,18 +313,15 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSessionReference firstDrmSessionReference =
         checkNotNull(
             drmSessionManager.preacquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     DrmSession firstDrmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(firstDrmSession);
     firstDrmSession.release(/* eventDispatcher= */ null);
 
@@ -343,9 +331,7 @@ public class DefaultDrmSessionManagerTest {
     DrmSession secondDrmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                secondFormatWithDrmInitData));
+                /* eventDispatcher= */ null, secondFormatWithDrmInitData));
     // The drmSessionManager had to release both it's internal keep-alive reference and the
     // reference represented by firstDrmSessionReference in order to acquire secondDrmSession.
     assertThat(firstDrmSession.getState()).isEqualTo(DrmSession.STATE_RELEASED);
@@ -372,12 +358,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession firstDrmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(firstDrmSession);
     firstDrmSession.release(/* eventDispatcher= */ null);
 
@@ -388,9 +373,7 @@ public class DefaultDrmSessionManagerTest {
     DrmSession secondDrmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     assertThat(secondDrmSession).isSameInstanceAs(firstDrmSession);
 
     // Let the timeout definitely expire, and check the session didn't get released.
@@ -422,12 +405,10 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
 
     DrmSessionReference sessionReference =
-        drmSessionManager.preacquireSession(
-            /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-            eventDispatcher,
-            FORMAT_WITH_DRM_INIT_DATA);
+        drmSessionManager.preacquireSession(eventDispatcher, FORMAT_WITH_DRM_INIT_DATA);
 
     // Wait for the key load event to propagate, indicating the pre-acquired session is in
     // STATE_OPENED_WITH_KEYS.
@@ -439,9 +420,7 @@ public class DefaultDrmSessionManagerTest {
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
 
     // Without idling the main/playback looper, we assert the session is already in OPENED_WITH_KEYS
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED_WITH_KEYS);
@@ -471,12 +450,10 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
 
     DrmSessionReference sessionReference =
-        drmSessionManager.preacquireSession(
-            /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-            /* eventDispatcher= */ null,
-            FORMAT_WITH_DRM_INIT_DATA);
+        drmSessionManager.preacquireSession(/* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA);
 
     // Release the pre-acquired reference before the underlying session has had a chance to be
     // constructed.
@@ -487,9 +464,7 @@ public class DefaultDrmSessionManagerTest {
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED);
 
     waitForOpenedWithKeys(drmSession);
@@ -511,12 +486,10 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
 
     DrmSessionReference sessionReference =
-        drmSessionManager.preacquireSession(
-            /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-            /* eventDispatcher= */ null,
-            FORMAT_WITH_DRM_INIT_DATA);
+        drmSessionManager.preacquireSession(/* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA);
 
     // Release the manager before the underlying session has had a chance to be constructed. This
     // will release all pre-acquired sessions.
@@ -531,9 +504,7 @@ public class DefaultDrmSessionManagerTest {
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED);
     waitForOpenedWithKeys(drmSession);
 
@@ -559,14 +530,13 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
 
     DefaultDrmSession drmSession =
         (DefaultDrmSession)
             checkNotNull(
                 drmSessionManager.acquireSession(
-                    /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                    /* eventDispatcher= */ null,
-                    FORMAT_WITH_DRM_INIT_DATA));
+                    /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
 
     assertThat(licenseServer.getReceivedSchemeDatas()).hasSize(1);
@@ -601,14 +571,13 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
 
     DefaultDrmSession drmSession =
         (DefaultDrmSession)
             checkNotNull(
                 drmSessionManager.acquireSession(
-                    /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                    /* eventDispatcher= */ null,
-                    FORMAT_WITH_DRM_INIT_DATA));
+                    /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     waitForOpenedWithKeys(drmSession);
 
     assertThat(licenseServer.getReceivedSchemeDatas()).hasSize(1);
@@ -646,12 +615,11 @@ public class DefaultDrmSessionManagerTest {
                 uuid -> new FakeExoMediaDrm.Builder().setProvisionsRequired(1).build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     // Confirm that opening the session threw NotProvisionedException (otherwise state would be
     // OPENED)
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENING);
@@ -680,12 +648,11 @@ public class DefaultDrmSessionManagerTest {
                         .build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED);
     waitForOpenedWithKeys(drmSession);
 
@@ -707,12 +674,11 @@ public class DefaultDrmSessionManagerTest {
                 uuid -> new FakeExoMediaDrm.Builder().setProvisionsRequired(2).build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     // Confirm that opening the session threw NotProvisionedException (otherwise state would be
     // OPENED)
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENING);
@@ -736,12 +702,11 @@ public class DefaultDrmSessionManagerTest {
                 DRM_SCHEME_UUID, uuid -> new FakeExoMediaDrm.Builder().build())
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENED);
     waitForOpenedWithKeys(drmSession);
 
@@ -763,12 +728,11 @@ public class DefaultDrmSessionManagerTest {
             .setSessionKeepaliveMs(C.TIME_UNSET)
             .build(/* mediaDrmCallback= */ licenseServer);
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     // Confirm that opening the session threw NotProvisionedException (otherwise state would be
     // OPENED)
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENING);
@@ -780,9 +744,7 @@ public class DefaultDrmSessionManagerTest {
     drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     // Confirm that opening the session threw NotProvisionedException (otherwise state would be
     // OPENED)
     assertThat(drmSession.getState()).isEqualTo(DrmSession.STATE_OPENING);
@@ -803,16 +765,12 @@ public class DefaultDrmSessionManagerTest {
         Exception.class,
         () ->
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     assertThrows(
         Exception.class,
         () ->
             drmSessionManager.preacquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
   }
 
   @Test
@@ -825,12 +783,11 @@ public class DefaultDrmSessionManagerTest {
             .build(/* mediaDrmCallback= */ licenseServer);
 
     drmSessionManager.prepare();
+    drmSessionManager.setPlayer(/* playbackLooper= */ Looper.myLooper(), PlayerId.UNSET);
     DrmSession drmSession =
         checkNotNull(
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     drmSessionManager.release();
 
     // The manager's prepareCount is now zero, but the drmSession is keeping it in a 'releasing'
@@ -839,16 +796,12 @@ public class DefaultDrmSessionManagerTest {
         Exception.class,
         () ->
             drmSessionManager.acquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
     assertThrows(
         Exception.class,
         () ->
             drmSessionManager.preacquireSession(
-                /* playbackLooper= */ checkNotNull(Looper.myLooper()),
-                /* eventDispatcher= */ null,
-                FORMAT_WITH_DRM_INIT_DATA));
+                /* eventDispatcher= */ null, FORMAT_WITH_DRM_INIT_DATA));
 
     drmSession.release(/* eventDispatcher= */ null);
   }